requirements.txt
delta-spark==3.1.0
pyspark==version 3.5.8
apache-airflow-providers-apache-spark